backend/requirements.txt
altgraph==0.17.4
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==2.0.12
Django==3.2.10
django-cors-headers==3.8.0
djangorestframework==3.12.4
gevent==23.9.1
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
psycopg2-binary==2.9.1
pytz==2023.3.post1
requests==2.26.0
sqlparse==0.4.4
typing-extensions==4.8.0
urllib3==1.26.18
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
pyjwt==2.3.1